Add lib directory to load path
These lines are standard for the default bundler output when creating a
new gem.

When I was having trouble getting the `Second` module loaded, my
instinct said to try:

```
require 'rack/throttle/second'
```

That didn't work because the `/lib` hadn't been added to the load path.
I don't know of any downside to doing so.